requirements.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.100.0
httpx==0.24.1
idna==3.4
pydantic==1.10.7
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1

PyMuPDF==1.23.7
Pillow==10.0.0
PyYAML==6.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6

networkx==3.2.1
longest-increasing-subsequence==0.1.7
numpy==1.23.5
pathos~=0.3.1
opencv-python==4.8.1.78
gunicorn==21.2.0
uvicorn==0.23.1

# cython==0.29.36
ujson==5.8.0
uvloop==0.17.0

boto3==1.28.17
aioboto3~=11.3.0
botocore~=1.31.17